+++ packages/connect/src/base/entry.spec.ts
@@ -101,7 +101,7 @@ describe('entry node functionality', function () {
     assert(uncheckedNodes[0].id.equals(peerStoreEntry.id), `id must match the generated one`)
     assert(uncheckedNodes[0].multiaddrs.length == peerStoreEntry.multiaddrs.length, `must not contain more multiaddrs`)
 
-    const usedRelays = entryNodes.getUsedRelays()
+    const usedRelays = entryNodes.getUsedRelayAddresses()
     assert(usedRelays == undefined || usedRelays.length == 0, `must not expose any internal addrs`)
 
     entryNodes.stop()
@@ -128,7 +128,7 @@ describe('entry node functionality', function () {
     const availablePublicNodes = entryNodes.getAvailabeEntryNodes()
     assert(availablePublicNodes.length == 0, `must remove node from public nodes`)
 
-    const usedRelays = entryNodes.getUsedRelays()
+    const usedRelays = entryNodes.getUsedRelayAddresses()
     assert(usedRelays == undefined || usedRelays.length == 0, `must not expose any internal addrs`)
 
     entryNodes.stop()
@@ -162,7 +162,7 @@ describe('entry node functionality', function () {
     assert(availableEntryNodes[0].id.equals(relay.id), `must contain correct peerId`)
     assert(availableEntryNodes[0].latency >= 0, `latency must be non-negative`)
 
-    const usedRelays = entryNodes.getUsedRelays()
+    const usedRelays = entryNodes.getUsedRelayAddresses()
     assert(usedRelays != undefined, `must expose relay addrs`)
     assert(usedRelays.length == 1, `must expose exactly one relay addrs`)
     assert(
@@ -207,7 +207,7 @@ describe('entry node functionality', function () {
 
     await Promise.all(relayNodes.map((relayNode) => relayNode[0]))
 
-    const usedRelays = entryNodes.getUsedRelays()
+    const usedRelays = entryNodes.getUsedRelayAddresses()
     assert(usedRelays != undefined, `must expose relay addresses`)
     assert(usedRelays.length == MAX_RELAYS_PER_NODE, `must expose ${MAX_RELAYS_PER_NODE} relay addresses`)
 
@@ -244,10 +244,15 @@ describe('entry node functionality', function () {
 
     entryNodes.uncheckedEntryNodes.push(newNode)
 
-    entryNodes.usedRelays.push(new Multiaddr(`/p2p/${relay.id.toB58String()}/p2p-circuit/p2p/${peerId.toB58String()}`))
+    let usedRelay = {
+      relayDirectAddress: new Multiaddr('/ip4/127.0.0.1/tcp/1234'),
+      ourCircuitAddress: new Multiaddr(`/p2p/${relay.id.toB58String()}/p2p-circuit/p2p/${peerId.toB58String()}`)
+    }
+
+    entryNodes.usedRelays.push(usedRelay)
 
     // Should have one unchecked node and one relay node
-    assert(entryNodes.getUsedRelays().length == 1)
+    assert(entryNodes.getUsedRelayAddresses().length == 1)
     assert(entryNodes.getUncheckedEntryNodes().length == 1)
 
     const connectPromise = once(newNodeListener, 'connected')
@@ -260,8 +265,8 @@ describe('entry node functionality', function () {
 
     assert(entryNodes.getAvailabeEntryNodes().length == 1)
 
-    const usedRelays = entryNodes.getUsedRelays()
-    assert(entryNodes.getUsedRelays().length == 1)
+    const usedRelays = entryNodes.getUsedRelayAddresses()
+    assert(entryNodes.getUsedRelayAddresses().length == 1)
 
     assert(
       usedRelays[0].equals(new Multiaddr(`/p2p/${newNode.id.toB58String()}/p2p-circuit/p2p/${peerId.toB58String()}`))
@@ -301,7 +306,7 @@ describe('entry node functionality', function () {
     assert(availableEntryNodes.length == 1)
     assert(availableEntryNodes[0].id.equals(relay.id))
 
-    const usedRelays = entryNodes.getUsedRelays()
+    const usedRelays = entryNodes.getUsedRelayAddresses()
     assert(usedRelays.length == 1)
     assert(
       usedRelays[0].equals(new Multiaddr(`/p2p/${relay.id.toB58String()}/p2p-circuit/p2p/${peerId.toB58String()}`))
@@ -329,7 +334,7 @@ describe('entry node functionality', function () {
 
     await entryNodes.updatePublicNodes()
 
-    const usedRelays = entryNodes.getUsedRelays()
+    const usedRelays = entryNodes.getUsedRelayAddresses()
     assert(usedRelays.length == 0)
 
     network.close()
@@ -343,8 +348,13 @@ describe('entry node functionality', function () {
 
     const relay = getPeerStoreEntry(`/ip4/127.0.0.1/tcp/1`)
 
+    let usedRelay = {
+      relayDirectAddress: new Multiaddr(`/ip4/127.0.0.1/tcp/1`),
+      ourCircuitAddress: new Multiaddr(`/p2p/${relay.id.toB58String()}/p2p-circuit/p2p/${peerId.toB58String()}`)
+    }
+
     entryNodes.availableEntryNodes.push({ ...relay, latency: 23 })
-    entryNodes.usedRelays.push(new Multiaddr(`/p2p/${relay.id.toB58String()}/p2p-circuit/p2p/${peerId.toB58String()}`))
+    entryNodes.usedRelays.push(usedRelay)
 
     entryNodes.once('listening', () =>
       assert.fail(`must not throw listening event if list of entry nodes has not changed`)
@@ -355,7 +365,7 @@ describe('entry node functionality', function () {
     const availableEntryNodes = entryNodes.getAvailabeEntryNodes()
     assert(availableEntryNodes.length == 0)
 
-    const usedRelays = entryNodes.getUsedRelays()
+    const usedRelays = entryNodes.getUsedRelayAddresses()
     assert(usedRelays.length == 0)
 
     entryNodes.stop()
@@ -431,5 +441,40 @@ describe('entry node functionality', function () {
     connectEmitter.removeAllListeners()
     entryNodes.stop()
     network.close()
+  })
+
+  it('do not contact nodes we are already connected to', async function () {
+    const entryNodes = new TestingEntryNodes(
+      peerId,
+      // Make sure that call is indeed asynchronous
+      (async () => new Promise((resolve) => setImmediate(resolve))) as any,
+      {}
+    )
+
+    const ma = new Multiaddr('/ip4/8.8.8.8/tcp/9091')
+
+    const peerStoreEntry = getPeerStoreEntry(ma.toString())
+
+    entryNodes.usedRelays.push({
+      relayDirectAddress: ma,
+      ourCircuitAddress: new Multiaddr(
+        `/p2p/${peerStoreEntry.id.toB58String()}/p2p-circuit/p2p/${peerId.toB58String()}`
+      )
+    })
+
+    entryNodes.start()
+
+    entryNodes.onNewRelay(peerStoreEntry)
+
+    const uncheckedNodes = entryNodes.getUncheckedEntryNodes()
+
+    assert(uncheckedNodes.length == 1, `Unchecked nodes must contain one entry`)
+    assert(uncheckedNodes[0].id.equals(peerStoreEntry.id), `id must match the generated one`)
+    assert(uncheckedNodes[0].multiaddrs.length == peerStoreEntry.multiaddrs.length, `must not contain more multiaddrs`)
+
+    const usedRelays = entryNodes.getUsedRelayAddresses()
+    assert(usedRelays.length == 1, `must not expose any internal addrs`)
+
+    entryNodes.stop()
   })
 })