+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 
 - New API v2 endpoint `/api/v2/node/stream/websockets` ([#3514](https://github.com/hoprnet/hoprnet/issues/3514))
+- Do not attempt to reconnect to relays we already have a connection to ([#3411](https://github.com/hoprnet/hoprnet/issues/3411))
 - New API v2 endpoint `/api/v2/node/peers` ([#3617](https://github.com/hoprnet/hoprnet/pull/3617))
 - Bug fix endpoint `/api/v2/channels/{peerId}` ([#3627](https://github.com/hoprnet/hoprnet/issues/3627))
 - Various bug fixes in `core`